+++ packages/classic/src/hashes/index.ts
@@ -3,7 +3,7 @@ export { ripemd160 } from './ripemd160.js';
 export { sha1 } from './sha1.js';
 export { sha224 } from './sha224.js';
 export { sha256 } from './sha256.js';
-export { sha3 } from './sha3.js';
+export { keccak512, sha3 } from './sha3.js';
 export { sha384 } from './sha384.js';
 export { sha512 } from './sha512.js';
 export { registerClassicHashes } from './register.js';