+++ packages/classic/src/hashes/md5.ts
@@ -1,4 +1,4 @@
 import { createHash } from './component.js';
-import { md5Hash } from './native.js';
+import { md5 as md5Digest } from 'js-md5';
 
-export const md5 = createHash('MD5', 64, 16, md5Hash);
+export const md5 = createHash('MD5', 64, 16, (input) => new Uint8Array(md5Digest.arrayBuffer(input)));