+++ packages/classic/src/hashes/sha224.ts
@@ -1,4 +1,4 @@
 import { createHash } from './component.js';
-import { sha224Hash } from './native.js';
+import { sha224 as sha224Digest } from 'js-sha256';
 
-export const sha224 = createHash('SHA224', 64, 28, sha224Hash);
+export const sha224 = createHash('SHA224', 64, 28, (input) => new Uint8Array(sha224Digest.arrayBuffer(input)));