+++ packages/classic/src/hashes/sha384.ts
@@ -1,4 +1,4 @@
 import { createHash } from './component.js';
-import { sha384Hash } from './native.js';
+import { sha384 as sha384Digest } from 'js-sha512';
 
-export const sha384 = createHash('SHA384', 128, 48, sha384Hash);
+export const sha384 = createHash('SHA384', 128, 48, (input) => new Uint8Array(sha384Digest.arrayBuffer(input)));