+++ packages/classic/src/hashes/sha512.ts
@@ -1,4 +1,4 @@
 import { createHash } from './component.js';
-import { sha512Hash } from './native.js';
+import { sha512 as sha512Digest } from 'js-sha512';
 
-export const sha512 = createHash('SHA512', 128, 64, sha512Hash);
+export const sha512 = createHash('SHA512', 128, 64, (input) => new Uint8Array(sha512Digest.arrayBuffer(input)));